+++ src/pages/calendar/Calendar.jsx
@@ -6,6 +6,7 @@ import CalendarSchedule from "./components/CalendarSchedule";
 import useModal from "../../shared/components/modal/useModal";
 import Modal from "./../../shared/components/modal/Modal";
 import ModalCalendar from "./components/ModalCalendar";
+import { useCalendar } from "./hooks/useCalendar";
 import { db, auth } from "../../shared/firebase";
 import { collection, addDoc, updateDoc, deleteDoc, doc } from "firebase/firestore";
 
@@ -37,10 +38,8 @@ const StyledCalendarWeek = styled.thead`
 
 const Calendar = () => {
   // 달력
-  const [currentDate, setCurrentDate] = useState(new Date());
-  const year = currentDate.getFullYear();
-  const month = currentDate.getMonth();
   const daysOfWeek = ["일", "월", "화", "수", "목", "금", "토"];
+  const { currentDate, weeks, handlePrevMonth, handleNextMonth } = useCalendar();
 
   // 모달
   const { isOpen, onOpen, onClose } = useModal();
@@ -51,65 +50,6 @@ const Calendar = () => {
   const [textAreaValue, setTextAreaValue] = useState("");
   const [selectedSchedule, setSelectedSchedule] = useState(null);
   const [isSubmitted, setIsSubmitted] = useState(false);
-
-  // 달력 생성
-  const generateCalendarDays = (year, month) => {
-    const firstDayOfMonth = new Date(year, month, 1).getDay(); // 첫 번째 날짜의 요일
-    const daysInMonth = new Date(year, month + 1, 0).getDate(); // 이번 달의 총 일 수
-    const prevMonthDays = new Date(year, month, 0).getDate(); // 이전 달의 총 일 수
-    const calendarDays = [];
-
-    // 첫 주 빈칸에 이전 달 날짜 추가
-    for (let i = firstDayOfMonth - 1; i >= 0; i--) {
-      calendarDays.push({
-        day: prevMonthDays - i,
-        isDisabled: true,
-        date: new Date(year, month - 1, prevMonthDays - i),
-      });
-    }
-
-    // 이번 달 날짜 추가
-    for (let i = 1; i <= daysInMonth; i++) {
-      calendarDays.push({
-        day: i,
-        isDisabled: false,
-        date: new Date(year, month, i),
-      });
-    }
-    
-    // 1일 부터 시작
-    let nextMonthDay = 1;
-
-    // 마지막 주 빈칸에 다음 달 날짜 추가
-    while (calendarDays.length % 7 !== 0) {
-      calendarDays.push({
-        day: nextMonthDay++,
-        isDisabled: true,
-        date: new Date(year, month + 1, nextMonthDay - 1),
-      });
-    }
-
-    const weeks = [];
-
-    // 7일씩 배열 생성
-    for (let i = 0; i < calendarDays.length; i += 7) {
-      weeks.push(calendarDays.slice(i, i + 7));
-    }
-
-    return weeks;
-  };
-
-  const weeks = generateCalendarDays(year, month);
-
-  // 이전 달로 이동
-  const handlePrevMonth = () => {
-    setCurrentDate(new Date(year, month - 1, 1));
-  };
-
-  // 다음 달로 이동
-  const handleNextMonth = () => {
-    setCurrentDate(new Date(year, month + 1, 1));
-  };
 
   const handleModalClose = () => {
     // 상태 초기화
@@ -220,8 +160,8 @@ const Calendar = () => {
       <PageTitle title="내 일정" />
       <StyledCalendarWrapper>
         <CalendarHeader
-          year={year}
-          month={month}
+          year={currentDate.getFullYear()}
+          month={currentDate.getMonth()}
           handlePrevMonth={handlePrevMonth}
           handleNextMonth={handleNextMonth}
           onOpen={onOpen}
